Removed Price and History and combined into single code base.
