+++ appsettings.json
@@ -7,7 +7,7 @@
     "Silver": "XAG",
     "Palladium": "XPD",
     "Platinum":  "XPT",
-    "Currency":  "USD",
+    "Currency": "USD",
     "MonthlyAllowance": "100"
   },
   "Logging": {
